+++ ui-core/src/index.ts
@@ -2,5 +2,5 @@ export { InvestigatorTermsOfUse } from './terms/InvestigatorTermsOfUse'
 export { ParticipantTermsOfUse } from './terms/ParticipantTermsOfUse'
 export { PrivacyPolicy } from './terms/PrivacyPolicy'
 
+export * from './types/forms'
 export * from './types/landingPageConfig'
-export * from './types/surveys'
